kernel: mtdsplit: okli: optional loader split
okli images always have the following structure:
* ELF kernel loader
* kernel uImage
* rootfs

Because the kernel loader is limited in size, the uImage can start
within the same erase block. The current version of the the uImage
splitter doesn't support partitions not starting at an erase block
boundary. This required either fixed partitions, which may limit the
supported kernel size, or manual splitting of the firmware partition.
Manual splitting into loader and firmware enabled mtdsplit support,
but requires the sysupgrade image to not include the loader, thus
complicating the device support directives.

This patch reimplements the original uImage splitter code to optionally
split off a third loader partition (for backwards compatibility), but
with a fixed kernel/rootfs partition order.