ath79: switch to plain loader for ar9344 CPE/WBS
Now that firmware partition starting with a plain ELF kernel loader can
be split automatically, convert the TP-Link AR9344-based CPE and WBS
devices from an OKLI loader to a plain kernel loader.

This reduces the size of the device images a bit (2kB for initramfs and
sysupgrade). More importantly it takes away the need to periodically
resize the kernel partition to allow for larger kernels.

Note that the firmware size in tplink-safeloader.c is one erase block
smaller than the one in the DTSI, due to some runtime space reclamation
happening on these devices.

Also rebases tplink-eap2x5 on the new tplink-safeloader-elf base device.

Signed-off-by: Sander Vanheule <sander@svanheule.net>