build: uImage: allow override of default arguments
For some build recipes, the argument to Build/uImage is used to sneak in
extra arguments for mkimage, whereas this appears to have been intended
to specificy the compression method.

Use a default compression value of 'none', and use the call arguments to
override the defaults provided in the Build/uImage recipe.  Only the -d
argument and the output file name cannot overriden.

Fixes all calls to Build/uImage to use the new notation.

Signed-off-by: Sander Vanheule <sander@svanheule.net>