kernel: add package for regulator userspace use
Device power outputs like PoE passthroughs are often enabled and
disabled via a GPIO. If this GPIO is linked to a fixed-regulator, an
additional userspace interface is required, since the regulator sysfs
interface is read-only.

Create a package for reg-userspace-consumer to allow userspace control
of GPIO controlled (and other) regulators.

Signed-off-by: Sander Vanheule <sander@svanheule.net>